+++ include/libopencm3/stm32/common/spi_common_v1_frf.h
@@ -29,36 +29,38 @@ specific memorymap.h header before including this header file.*/
 /** @cond */
 #ifdef LIBOPENCM3_SPI_H
 /** @endcond */
-#ifndef LIBOPENCM3_SPI_COMMON_L1F124_H
-#define LIBOPENCM3_SPI_COMMON_L1F124_H
+#pragma once
 
 /**@{*/
 
-#include <libopencm3/stm32/common/spi_common_all.h>
+#include <libopencm3/stm32/common/spi_common_v1.h>
 
-/*
- * This file extends the common STM32 version with definitions only
- * applicable to the STM32L1/F1/2/4 series of devices.
- */
+/* --- SPI_CR2 values ------------------------------------------------------ */
 
-/* DFF: Data frame format */
-/****************************************************************************/
-/** @defgroup spi_dff SPI data frame format
-@ingroup spi_defines
+/* FRF: Frame format */
+/* Note: Not used in I2S mode. */
+#define SPI_CR2_FRF				(1 << 4)
+#define SPI_CR2_FRF_MOTOROLA_MODE	(0 << 4)
+#define SPI_CR2_FRF_TI_MODE		(1 << 4)
 
-@{*/
+/* --- SPI_SR values ------------------------------------------------------- */
 
-#define SPI_CR1_DFF_8BIT        (0 << 11)
-#define SPI_CR1_DFF_16BIT       (1 << 11)
+/* FRE / TIFRFE: TI frame format error */
+#define SPI_SR_TIFRFE			(1 << 8) //F2
+#define SPI_SR_FRE				(1 << 8) //others
 
-/**@}*/
+/* --- Function prototypes ------------------------------------------------- */
 
-#define SPI_CR1_DFF         (1 << 11)
+BEGIN_DECLS
 
-#endif
+void spi_set_frf_ti(uint32_t spi);
+void spi_set_frf_motorola(uint32_t spi);
+
+END_DECLS
+
 /** @cond */
 #else
-#warning "spi_common_l1f124.h should not be included explicitly, only via spi.h"
+#warning "spi_common_v1_frf.h should not be included explicitly, only via spi.h"
 #endif
 /** @endcond */
 /**@}*/
