+++ include/libopencm3/stm32/l0/spi.h
@@ -1,14 +1,14 @@
 /** @defgroup spi_defines SPI Defines
- *
- * @brief <b>Defined Constants and Types for the STM32F3xx SPI</b>
- *
- * @ingroup STM32F3xx_defines
- *
- * @version 1.0.0
- *
- * @date 5 December 2012
- *
- * LGPL License Terms @ref lgpl_license
+
+@brief <b>Defined Constants and Types for the STM32L0xx SPI</b>
+
+@ingroup STM32L0xx_defines
+
+@version 1.0.0
+
+@date 20 January 2017
+
+LGPL License Terms @ref lgpl_license
  */
 
 /*
@@ -31,6 +31,7 @@
 #ifndef LIBOPENCM3_SPI_H
 #define LIBOPENCM3_SPI_H
 
-#include <libopencm3/stm32/common/spi_common_v2.h>
+#include <libopencm3/stm32/common/spi_common_v1_frf.h>
 
 #endif
+
