+++ include/libopencm3/stm32/spi.h
@@ -30,6 +30,8 @@
 #       include <libopencm3/stm32/f3/spi.h>
 #elif defined(STM32F4)
 #       include <libopencm3/stm32/f4/spi.h>
+#elif defined(STM32L0)
+#       include <libopencm3/stm32/l0/spi.h>
 #elif defined(STM32L1)
 #       include <libopencm3/stm32/l1/spi.h>
 #elif defined(STM32L4)
