+++ lib/stm32/f0/Makefile
@@ -41,9 +41,10 @@ OBJS		= can.o flash.o rcc.o usart.o dma.o rtc.o comparator.o crc.o \
 
 OBJS            += gpio_common_all.o gpio_common_f0234.o crc_common_all.o crc_v2.o \
                    pwr_common_v1.o iwdg_common_all.o rtc_common_l1f024.o \
-                   dma_common_l1f013.o exti_common_all.o spi_common_all.o \
-		   spi_common_f03.o flash_common_f01.o dac_common_all.o \
-		   timer_common_all.o timer_common_f0234.o rcc_common_all.o
+                   dma_common_l1f013.o exti_common_all.o \
+				   flash_common_f01.o dac_common_all.o \
+				   timer_common_all.o timer_common_f0234.o rcc_common_all.o
+
 OBJS		+= adc_common_v2.o
 OBJS		+= crs_common_all.o
 OBJS		+= usart_common_all.o usart_common_v2.o
@@ -51,6 +52,7 @@ OBJS		+= i2c_common_v2.o
 
 OBJS		+= usb.o usb_control.o usb_standard.o
 OBJS		+= st_usbfs_core.o st_usbfs_v2.o
+OBJS		+= spi_common_all.o spi_common_v2.o
 
 VPATH += ../../usb:../:../../cm3:../common
