+++ lib/stm32/f2/Makefile
@@ -40,12 +40,13 @@ OBJS		= gpio.o rcc.o desig.o
 
 OBJS            += crc_common_all.o dac_common_all.o dma_common_f24.o \
                    gpio_common_all.o gpio_common_f0234.o i2c_common_v1.o \
-                   iwdg_common_all.o rtc_common_l1f024.o spi_common_all.o \
-                   spi_common_l1f124.o timer_common_all.o timer_common_f0234.o \
+                   iwdg_common_all.o rtc_common_l1f024.o \
+                   timer_common_all.o timer_common_f0234.o \
 		   timer_common_f24.o usart_common_all.o usart_common_f124.o \
 		   flash_common_f234.o flash_common_f24.o hash_common_f24.o \
 		   crypto_common_f24.o exti_common_all.o rcc_common_all.o
 OBJS		+= rng_common_v1.o
+OBJS            += spi_common_all.o spi_common_v1.o spi_common_v1_frf.o
 
 OBJS            += usb.o usb_standard.o usb_control.o usb_dwc_common.o \
                    usb_f107.o usb_f207.o usb_msc.o
