+++ lib/stm32/f4/Makefile
@@ -45,12 +45,13 @@ OBJS		= adc.o adc_common_v1.o can.o desig.o gpio.o pwr.o rcc.o \
 OBJS            += crc_common_all.o dac_common_all.o dma_common_f24.o \
                    gpio_common_all.o gpio_common_f0234.o i2c_common_v1.o \
                    iwdg_common_all.o pwr_common_v1.o rtc_common_l1f024.o \
-                   spi_common_all.o spi_common_l1f124.o timer_common_all.o \
+                   timer_common_all.o \
                    timer_common_f0234.o timer_common_f24.o usart_common_all.o \
 		   usart_common_f124.o flash_common_f234.o flash_common_f24.o \
 		   hash_common_f24.o crypto_common_f24.o exti_common_all.o \
 		   rcc_common_all.o
 OBJS		+= rng_common_v1.o
+OBJS		+= spi_common_all.o spi_common_v1.o spi_common_v1_frf.o
 
 OBJS            += usb.o usb_standard.o usb_control.o usb_dwc_common.o \
 		   usb_f107.o usb_f207.o usb_msc.o
