+++ lib/stm32/l1/Makefile
@@ -42,7 +42,8 @@ OBJS		+= flash_common_l01.o
 OBJS		+= gpio_common_all.o gpio_common_f0234.o
 OBJS		+= i2c_common_v1.o iwdg_common_all.o
 OBJS		+= pwr_common_v1.o pwr_common_v2.o rtc_common_l1f024.o
-OBJS		+= spi_common_all.o spi_common_l1f124.o timer_common_all.o
+OBJS		+= spi_common_all.o spi_common_v1.o spi_common_v1_frf.o
+OBJS		+= timer_common_all.o
 OBJS		+= usart_common_all.o usart_common_f124.o
 OBJS		+= exti_common_all.o
 OBJS		+= rcc_common_all.o
